Delete status field, add docid field.
The status field isn't required since we simply won't insert into the
DB if the publish doesn't succeed. Still need to figure out how to
deal with content updates.

On successful submit, it returns the CouchDB doc_ID so we can store it
in the DB and may be able to use it for making assertions about LR
docs that refer to old versions.